backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.7.4
pandas==2.1.4
nba_api==1.4.1
sentence-transformers==2.2.2
huggingface-hub==0.17.3
numpy==1.26.2
python-dotenv==1.0.0
scikit-learn==1.3.2
google-generativeai==0.8.4
google-generativeai==0.8.4
